requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0

# PostgreSQL + pgvector
psycopg[binary]>=3.2.0
pgvector>=0.3.0

# LanceDB
lancedb>=0.17.0
pyarrow>=18.0.0

# Ollama client
ollama>=0.4.0

# Configuration
pydantic>=2.10.0
pydantic-settings>=2.7.0
pyyaml>=6.0.0

# Transcript downloading
yt-dlp>=2025.1.0

# Text processing
webvtt-py>=0.5.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.28.0
